refactor(pipeline-email): fold facts.ts into stages/
